Fail after a (configurable) timeout
If we make it more than say 5 minutes we've probably gotten past
the point where GN would complain and any failures are likely to
require actual dev effort.

Just die with an error at that point to save on electricity.

+++ iterate-over-ebuild.sh
@@ -9,6 +9,7 @@ package="${1%.ebuild}"
 tmpfile=$(mktemp)
 iter=0
 added=()
+timeout_secs=300
 
 # Trap for Ctrl+C
 trap 'cleanup' INT
@@ -20,6 +21,7 @@ cleanup() {
 }
 
 while true; do
+  start_time=$(date +%s)
   libs=()
   echo "[$(date)]: Processing $package; iteration $((++iter))"
   echo "So far, we've added:"
@@ -54,6 +56,15 @@ while true; do
     rm "$tmpfile"
     break
   fi
+
+  end_time=$(date +%s)
+  elapsed_time=$((end_time - start_time))
+  if [ $elapsed_time -gt $timeout_secs ]; then
+    echo "[$(date)]: Ebuild execution took longer than the timeout. This is likely a build failure that requires patching. Exiting."
+    echo "$tmpfile" for this iteration\'s logs.
+    exit 1
+  fi
+
   # Start with a clean slate for the next iteration
   rm "$tmpfile"
 done
